requirements.txt
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
Flask==2.0.1
gunicorn==20.1.0
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
PyYAML==5.4.1
requests==2.26.0
urllib3==1.26.6
Werkzeug==2.0.1
